# [fips-9-compliant] perf: Disallow mis-matched inherited group reads