@thefossguy-ciq 2025-08-04
> upstream-diff [...]
> [...]
> Also, add an upstream patch 28a6c6e ("perf/core: Fix potential NULL deref")
> which fixes a NULL pointer deref issue in the existing CVE fix.

I don't know what's the consensus on the above addition to the `upstream-diff`. I'm indifferent but want to know @PlaidCat's thoughts.

The rest looks good though!